# docs/swagger.yaml
basePath: /
definitions:
  history-api_internal_dtos_request.AddProjectMemberDto:
    properties:
      role:
        enum:
        - EDITOR
        - VIEWER
        type: string
      user_id:
        type: string
    required:
    - role
    - user_id
    type: object
  history-api_internal_dtos_request.ChangeOwnerDto:
    properties:
      new_owner_id:
        type: string
    required:
    - new_owner_id
    type: object
...
  history-api_internal_dtos_request.CreateProjectDto:
    properties:
      description:
        type: string
      status:
        enum:
        - PRIVATE
        - PUBLIC
        - ARCHIVE
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - title
    type: object
...
  history-api_internal_dtos_request.UpdateProjectDto:
    properties:
      description:
        type: string
      status:
        enum:
        - PRIVATE
        - PUBLIC
        - ARCHIVE
        type: string
      title:
        maxLength: 255
        type: string
    type: object
  history-api_internal_dtos_request.UpdateProjectMemberDto:
    properties:
      role:
        enum:
        - EDITOR
        - VIEWER
        type: string
    required:
    - role
    type: object
...
  history-api_internal_dtos_response.CommonResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  history-api_internal_dtos_response.PaginatedResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/history-api_internal_dtos_response.PaginationMeta'
      status:
        type: boolean
    type: object
  history-api_internal_dtos_response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for History API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: History API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Search and filter projects with pagination
      parameters:
      - in: query
        name: created_from
        type: string
      - in: query
        name: created_to
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maxLength: 200
        name: search
        type: string
      - enum:
        - created_at
        - updated_at
        - title
        in: query
        name: sort
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: statuses
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: user_ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Search projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a project for the current authenticated user
      parameters:
      - description: Project Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.CreateProjectDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Retrieve project details by specific ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Update project properties (Title, Description, Status)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.UpdateProjectDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - Projects
  /projects/{id}/change-owner:
    put:
      consumes:
      - application/json
      description: Transfer project ownership to an existing member. Only the current
        owner can do this.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: New Owner Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.ChangeOwnerDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Transfer project ownership
      tags:
      - Projects
...
  /projects/{id}/members:
    post:
      consumes:
      - application/json
      description: Invite a user to the project with a specific role (EDITOR or VIEWER).
        Only project owner can do this.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Member Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.AddProjectMemberDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Add a member to project
      tags:
      - Projects
  /projects/{id}/members/{userId}:
    delete:
      consumes:
      - application/json
      description: Remove a user from the project. Only project owner can do this.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Member User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Remove a member from project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Change a member's role in the project. Only project owner can do
        this.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Member User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.UpdateProjectMemberDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - Projects
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"